pyproject.toml
[project]
name = "hestia"
version = "0.1.0"
description = "Your Personal Server Hearth"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.41",
]
